Refactor into playlist module
