+++ client/src/playlist/components/AlbumPreview.vue
