+++ client/src/playlist/routes.js
@@ -1,0 +1,9 @@
+const routes = [
+  {
+    path: "/playlist",
+    name: "playlist",
+    component: () => import("./views/PlaylistView.vue"),
+  },
+];
+
+export default routes;
