+++ client/src/views/music/MusicView.vue
@@ -8,7 +8,7 @@
 <script>
 import BreadCrumbs from "@/components/music/BreadCrumbs.vue";
 import { mapStores } from "pinia";
-import { usePlaylistStore } from "@/stores/playlist";
+import { usePlaylistStore } from "@/playlist/store";
 
 export default {
   components: { BreadCrumbs },
